requirements.txt
gradio
transformers
torch
qdrant-client
pymupdf
pillow